Part5: pr#56 metric (#86)
* add metric

* Part6: pr#60 ratelimit (#87)

* small fix (#88)
